requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
werkzeug
flask-cors
pymysql
cryptography
